Calculate v2.1 Exception

+++ Calculate.java
@@ -1,16 +1,16 @@
 public class Calculate {
     private double result;
 
-    public void summa(int first, int second)
+    public void summa(double first, double second)
     {
         result = first + second;
     }
-    public void vichitanie(int first, int second)
+    public void vichitanie(double first, double second)
     {
         result = first - second;
     }
 
-    public void proizv(int first, int second)
+    public void proizv(double first, double second)
     {
         result = first * second;
     }
@@ -20,7 +20,7 @@ public class Calculate {
         result = (first / second) + (first % second);
     }
 
-    public void stepen(int first, int second)
+    public void stepen(double first, double second)
     {
         result = (int) Math.pow(first,second);
     }
